#268 rename: Relocate User related components into mobile and desktop
